refactor(mfa): split the package into manager/totp/webauthn files
Rename mfa.go into manager.go (Manager, Provider, chain and step logic),
totp.go (TOTP provider and endpoint) and webauthn.go, with matching
totp_test.go and webauthn_test.go, instead of a mismatched manager_test.go.
